Move the location reconstruction to a new file
so that we can:
- reuse it for the other segmentation
- make it more complex, as necessary
