+++ emission/analysis/intake/segmentation/section_segmentation_methods/smoothed_high_confidence_with_visit_transitions.py
@@ -7,12 +7,12 @@ from future import standard_library
 standard_library.install_aliases()
 from builtins import *
 import logging
-import pandas as pd
 
 # Our imports
 import emission.analysis.intake.segmentation.section_segmentation_methods.smoothed_high_confidence_motion as eaisms
 import emission.core.wrapper.motionactivity as ecwm
 import emission.core.wrapper.location as ecwl
+import emission.analysis.intake.segmentation.section_segmentation_methods.reconstruct_locations as eaissr
 
 class SmoothedHighConfidenceMotionWithVisitTransitions(eaisms.SmoothedHighConfidenceMotion):
     def create_unknown_section(self, location_points_df):
@@ -121,11 +121,13 @@ class SmoothedHighConfidenceMotionWithVisitTransitions(eaisms.SmoothedHighConfid
             resampled_sec_df = self.filter_points_for_range(self.resampled_loc_df,
                                         start_motion, end_motion)
 
-            if len(raw_section_df) and len(resampled_sec_df) == 0:
-                logging.info("Found no location points between %s and %s" % (start_motion, end_motion))
+            if len(raw_section_df) <=1 and len(unfiltered_section_df) <= 1:
+                logging.info("Found %d filtered points and %d unfiltered points between %s and %s for type %s, skipping..." % (len(raw_section_df), len(unfiltered_section_df),
+                                      start_motion.fmt_time, end_motion.fmt_time,
+                                      start_motion.type))
             else:
-                section_start_loc = get_matched_location(start_motion, raw_section_df, resampled_sec_df, 0, timeseries)
-                section_end_loc = get_matched_location(end_motion, raw_section_df, resampled_sec_df, -1, timeseries)
+                section_start_loc = eaissr.get_matched_location(start_motion, raw_section_df, resampled_sec_df, 0, timeseries)
+                section_end_loc = eaissr.get_matched_location(end_motion, raw_section_df, resampled_sec_df, -1, timeseries)
                 logging.debug("section start point = %s, section end point = %s" %
                               (ecwl.Location(section_start_loc), ecwl.Location(section_end_loc)))
                 section_list.append((section_start_loc, section_end_loc, start_motion.type))
@@ -134,41 +136,3 @@ class SmoothedHighConfidenceMotionWithVisitTransitions(eaisms.SmoothedHighConfid
             # will mark it as a transition
         return section_list
 
-def get_matched_location(motion, raw_section_df, resampled_sec_df, index, timeseries):
-    if len(raw_section_df) > 0:
-        orig_matched_point = raw_section_df.iloc[index]
-    else:
-        orig_matched_point = None
-
-    # If real points are fresh enough (within 30 secs of the motion
-    # activity start, use them), otherwise insert and use
-    # interpolated points
-    FRESHNESS_CHECK_THRESHOLD = 30
-    if orig_matched_point is not None and abs(orig_matched_point.ts - motion.ts) < FRESHNESS_CHECK_THRESHOLD:
-        logging.debug("matched_point %s is %d (within %d) secs from motion %s, retaining" % 
-            (orig_matched_point.fmt_time, abs(orig_matched_point.ts - motion.ts), FRESHNESS_CHECK_THRESHOLD, motion.fmt_time))
-        matched_point = orig_matched_point
-    else:
-        matched_point = resampled_sec_df.iloc[index]
-
-        if orig_matched_point is not None:
-            logging.debug("matched_point %s is %d secs from motion %s, using resampled location %s" % 
-                (orig_matched_point.fmt_time, abs(orig_matched_point.ts - motion.ts),
-                     motion.fmt_time, matched_point.fmt_time))
-            user_id = orig_matched_point.user_id
-        else:
-            logging.debug("matched_point %s for motion %s, using resampled location %s" % 
-                (orig_matched_point, motion.fmt_time, matched_point.fmt_time))
-            user_id = timeseries.user_id
-
-        # Make sure to insert the resampled point, so that it will exist if/when
-        # we look it up later
-        # TODO: Set this to the right version
-        matched_point = matched_point.append(pd.Series({"filter": "distance"}))
-        # Mark newly inserted location points as special so that we can move them
-        # later if we need to
-        matched_point = matched_point.append(pd.Series({"inserted": True}))
-        new_id = timeseries.insert_data(user_id, "background/filtered_location", ecwl.Location(matched_point))
-        matched_point = matched_point.append(pd.Series({"_id": new_id}))
-
-    return matched_point
